moved label_cleaner.py and cleaner.py to core
