+++ core/src/autogluon/core/data/__init__.py
@@ -1,0 +1,1 @@
+from .label_cleaner import LabelCleaner
